+++ src/test-ranking.cpp
@@ -328,7 +328,7 @@ int main(int argc, char *argv[])
     std::cout << "Single thread         : " << (singleThread ? "true" : "false") << std::endl;
 
     std::cout << std::fixed << std::setprecision(2);
-    const int numThreads = std::thread::hardware_concurrency();
+    const size_t numThreads = std::thread::hardware_concurrency();
     std::cout << "Number of threads     : " << numThreads << std::endl;
 
     if (!tieCorrection)
